Moved all ontology based queries and processing to their own directory
