+++ TapControl/tapcontrol/data/gen_nodes.json
@@ -4,16 +4,16 @@
       "x": 200,
       "y": 300,
       "connections": [
-        "632",
-        "650"
+        "650",
+        "632"
       ]
     },
     "650": {
       "x": 200,
       "y": 350,
       "connections": [
-        "RG60",
-        "SourceBus"
+        "SourceBus",
+        "RG60"
       ]
     },
     "611": {
@@ -27,8 +27,8 @@
       "x": 200,
       "y": 250,
       "connections": [
-        "RG60",
         "633",
+        "RG60",
         "670",
         "645"
       ]
@@ -52,8 +52,8 @@
       "x": 100,
       "y": 250,
       "connections": [
-        "632",
-        "646"
+        "646",
+        "632"
       ]
     },
     "646": {
@@ -83,9 +83,9 @@
       "y": 100,
       "connections": [
         "684",
+        "692",
         "670",
-        "680",
-        "692"
+        "680"
       ]
     },
     "675": {
@@ -106,17 +106,17 @@
       "x": 100,
       "y": 100,
       "connections": [
-        "652",
+        "611",
         "671",
-        "611"
+        "652"
       ]
     },
     "692": {
       "x": 250,
       "y": 100,
       "connections": [
-        "671",
-        "675"
+        "675",
+        "671"
       ]
     },
     "SourceBus": {
