--- TapControl/tapcontrol/data/test.json
@@ -1,1 +1,0 @@
-{'611': {'x': 0, 'y': 100, 'connections': ['684']}, '632': {'x': 200, 'y': 250, 'connections': ['633', '645', '670', 'RG60']}, '633': {'x': 350, 'y': 250, 'connections': ['632', '634']}, '634': {'x': 400, 'y': 250, 'connections': ['633']}, '645': {'x': 100, 'y': 250, 'connections': ['632', '646']}, '646': {'x': 0, 'y': 250, 'connections': ['645']}, '650': {'x': 200, 'y': 350, 'connections': ['SourceBus', 'RG60', 'RG60', 'RG60']}, '652': {'x': 100, 'y': 0, 'connections': ['684']}, '670': {'x': 200, 'y': 200, 'connections': ['632', '671']}, '671': {'x': 200, 'y': 100, 'connections': ['670', '680', '684', '692']}, '675': {'x': 400, 'y': 100, 'connections': ['692']}, '680': {'x': 200, 'y': 0, 'connections': ['671']}, '684': {'x': 100, 'y': 100, 'connections': ['671', '611', '652']}, '692': {'x': 250, 'y': 100, 'connections': ['675', '671']}, 'RG60': {'x': 200, 'y': 300, 'connections': ['632', '650', '650', '650']}, 'SourceBus': {'x': 200, 'y': 400, 'connections': ['650']}}
